move table functions to table service
